requirements.txt
openai~=1.76.0
python-dotenv~=1.1.0
PyPDF2~=3.0.1
numpy~=2.2.5
scikit-learn~=1.6.1